rename trainer modules, drop `_mixin` (#571)
* rename trainer modules, drop _mixin

* fix imports
