+++ pytorch_lightning/trainer/auto_mix_precision.py
@@ -18,7 +18,7 @@ class TrainerAMPMixin(ABC):
 
         if use_amp and not APEX_AVAILABLE:  # pragma: no cover
             msg = """
-            You set use_amp=True but do not have apex installed.
+            You set `use_amp=True` but do not have apex installed.
             Install apex first using this guide and rerun with use_amp=True:
             https://github.com/NVIDIA/apex#linux
