+++ pytorch_lightning/trainer/callback_config.py
