+++ pytorch_lightning/trainer/trainer.py
@@ -13,21 +13,21 @@ import torch.multiprocessing as mp
 import tqdm
 from torch.optim.optimizer import Optimizer
 
-from pytorch_lightning.trainer.amp_mixin import TrainerAMPMixin
-from pytorch_lightning.trainer.callback_config_mixin import TrainerCallbackConfigMixin
-from pytorch_lightning.trainer.data_loading_mixin import TrainerDataLoadingMixin
-from pytorch_lightning.trainer.ddp_mixin import TrainerDDPMixin
-from pytorch_lightning.trainer.dp_mixin import TrainerDPMixin
-from pytorch_lightning.trainer.dp_mixin import (
+from pytorch_lightning.trainer.auto_mix_precision import TrainerAMPMixin
+from pytorch_lightning.trainer.callback_config import TrainerCallbackConfigMixin
+from pytorch_lightning.trainer.data_loading import TrainerDataLoadingMixin
+from pytorch_lightning.trainer.distrib_data_parallel import TrainerDDPMixin
+from pytorch_lightning.trainer.distrib_parts import (
+    TrainerDPMixin,
     parse_gpu_ids,
     determine_root_gpu_device
 )
-from pytorch_lightning.trainer.evaluation_loop_mixin import TrainerEvaluationLoopMixin
-from pytorch_lightning.trainer.logging_mixin import TrainerLoggingMixin
-from pytorch_lightning.trainer.model_hooks_mixin import TrainerModelHooksMixin
-from pytorch_lightning.trainer.train_loop_mixin import TrainerTrainLoopMixin
+from pytorch_lightning.trainer.evaluation_loop import TrainerEvaluationLoopMixin
+from pytorch_lightning.trainer.logging import TrainerLoggingMixin
+from pytorch_lightning.trainer.model_hooks import TrainerModelHooksMixin
+from pytorch_lightning.trainer.training_loop import TrainerTrainLoopMixin
 from pytorch_lightning.trainer.trainer_io import TrainerIOMixin
-from pytorch_lightning.trainer.training_tricks_mixin import TrainerTrainingTricksMixin
+from pytorch_lightning.trainer.training_tricks import TrainerTrainingTricksMixin
 from pytorch_lightning.utilities.debugging import MisconfigurationException
 
 try:
